Add Asciidoc version of text.
As a reference, I used a scanned copy from the Internet Archive:

https://archive.org/stream/ladysusanwatson00austgoog#page/n6/mode/2up

As such, the Asciidoc version includes the preface from the scanned
copy.

Additional changes:

Capitilization of first words of each chapter.
Consistency with periods after signatures.
Consistency in formatting of signatures.

Conversion to Asciidoc revealed the following shortcomings of
Asciidoc (or of my knowledge of Asciidoc):
- Indicating that small caps should be used for a segment of text.
- Centering or right aligning text.